Created the MVC template. Cleaned the Code.
